+++ R/rmd_select.R
@@ -13,6 +13,8 @@
 #' @param ... One or more unquoted expressions separated by commas. Chunk labels can be used
 #' as if they were positions in the data frame, so expressions like x:y can be used to
 #' select a range of nodes.
+#' @param keep_yaml Logical, whether to automatically include YAML nodes in the selection.
+#' If `TRUE` (default), equivalent to including `has_type("rmd_yaml")` in the selection.
 #'
 #' @return Returns a subset Rmd object (either `rmd_ast` or `rmd_tibble` depending on input).
 #'
@@ -29,27 +31,41 @@
 #'
 #' @export
 #'
-rmd_select = function(x, ...) {
+rmd_select = function(x, ..., keep_yaml = TRUE) {
   UseMethod("rmd_select")
 }
 
 #' @exportS3Method
-rmd_select.default = function(x, ...) {
+rmd_select.default = function(x, ..., keep_yaml = TRUE) {
   stop("Unsupported type")
 }
 
 
-rmd_select_impl = function(x, ...) {
+rmd_select_impl = function(x, ..., keep_yaml = TRUE) {
   checkmate::assert_class(x, "rmd_ast")
+  checkmate::assert_logical(keep_yaml, len = 1, any.missing = FALSE)
 
   nodes = setNames(x@nodes, rmd_node_label(x))
-  loc = tidyselect::eval_select(rlang::expr(c(...)), nodes)
+  
+  # Check if we have any selection expressions
+  dots = rlang::enquos(...)
+  if (length(dots) > 0) {
+    loc = tidyselect::eval_select(rlang::expr(c(...)), nodes)
+  } else {
+    loc = integer(0)
+  }
+  
+  # Add YAML nodes if keep_yaml is TRUE
+  if (keep_yaml) {
+    yaml_indices = which(rmd_node_type(x) == "rmd_yaml")
+    loc = c(loc, yaml_indices)
+  }
 
-  sort(loc) # maintain original order
+  sort(unique(loc)) # maintain original order and remove duplicates
 }
 
 #' @exportS3Method
-rmd_select.rmd_tibble = function(x, ...) {
+rmd_select.rmd_tibble = function(x, ..., keep_yaml = TRUE) {
   sec_cols = names(x)[grepl("^sec_h", names(x))]
   bad_cols = sec_cols[!sec_cols %in% paste0("sec_h", 1:6)]
 
@@ -57,7 +73,7 @@ rmd_select.rmd_tibble = function(x, ...) {
     stop("The following columns must be renamed: ", bad_cols, call. = FALSE)
 
   x_ast = as_ast(x)
-  loc = rmd_select_impl(x_ast, ...)
+  loc = rmd_select_impl(x_ast, ..., keep_yaml = keep_yaml)
 
   x = x[loc,]
 
@@ -71,20 +87,20 @@ rmd_select.rmd_tibble = function(x, ...) {
 }
 
 #' @exportS3Method
-rmd_select.rmd_ast = function(x, ...) {
-  loc = rmd_select_impl(x, ...)
+rmd_select.rmd_ast = function(x, ..., keep_yaml = TRUE) {
+  loc = rmd_select_impl(x, ..., keep_yaml = keep_yaml)
   x@nodes = x@nodes[loc]
 
   x
 }
 
 #' @exportS3Method
-rmd_select.rmd_collection = function(x, ...) {
-  x$ast = purrr::map(x$ast, rmd_select, ...)
+rmd_select.rmd_collection = function(x, ..., keep_yaml = TRUE) {
+  x$ast = purrr::map(x$ast, rmd_select, ..., keep_yaml = keep_yaml)
   x
 }
 
 #' @exportS3Method
-rmd_select.qmd_collection = function(x, ...) {
-  rmd_select.rmd_collection(x, ...)
+rmd_select.qmd_collection = function(x, ..., keep_yaml = TRUE) {
+  rmd_select.rmd_collection(x, ..., keep_yaml = keep_yaml)
 }